Abandon tree-sitter grammar
There were some notable issues with users installing v2 of this package
because of the new tree-sitter dependencies and the build tools required
to compile them, which not everyone has.

I may revisit this in the future, but for now, I'm abandoning the
tree-sitter grammar and preserving the newly-added merchzone.txt support
via a separate first-mate grammar (which has JS disabled because it
would otherwise break the syntax highlighting due to the unclosed blocks
common to merchzone.txt files).